requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-pptx==1.0.2
anthropic==0.34.2
python-multipart==0.0.9
groq==0.9.0